update for v2.0 of Polynomials
