+++ Project.toml
@@ -4,27 +4,23 @@ authors = ["jverzani <jverzani@gmail.com>"]
 version = "0.2.0"
 
 [deps]
-FastGaussQuadrature = "442a2c76-b920-505d-bb47-c5924d526838"
-FastTransforms = "057dd010-8810-581a-b7be-e3fc3b93f78c"
 HypergeometricFunctions = "34004b35-14d8-5ef3-9330-4cdb6864b03a"
+Intervals = "d8418881-c3e1-53bb-8760-2df7ec849ed5"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Memoize = "c03570c3-d221-55d1-a50c-7939bbd78826"
 Polynomials = "f27b6e38-b328-58d1-80ce-0feddd5e7a45"
-Intervals = "d8418881-c3e1-53bb-8760-2df7ec849ed5"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 
 [compat]
+HypergeometricFunctions = "0.2, 0.3"
+Intervals = "0.5, 1"
 Memoize = "0.4"
 Polynomials = "2"
 QuadGK = "1,2"
-Intervals = "1.5"
 SpecialFunctions = "0.09, 0.10, 1"
-FastGaussQuadrature = "0.4"
-FastTransforms="0.9, 0.10, 0.11, 0.12"
-HypergeometricFunctions="0.2, 0.3"
 julia = "1.0"
-
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
