+++ src/Orthogonal/Jacobi.jl
@@ -88,8 +88,8 @@ function jacobi_eval(α, β, n, z)
 end
 
 
-gauss_nodes_weights(p::Type{P}, n) where {α, β, P <: Jacobi{α, β}} =
-    FastGaussQuadrature.gaussjacobi(n, α, β)
+# gauss_nodes_weights(p::Type{P}, n) where {α, β, P <: Jacobi{α, β}} =
+#     FastGaussQuadrature.gaussjacobi(n, α, β)
 
 function classical_hypergeometric(::Type{<:Jacobi{α, β}}, n, x) where {α,β}
