+++ src/Orthogonal/connection.jl
@@ -85,134 +85,134 @@ function _convert_cop(::Type{Q}, p::P) where {P <: ConvertibleTypes,
     ⟒(Q){R,X}(as)
 end
 
-## Use FastTransform for conversion, as  possible
-## FastTransforms.kind2string.(0:15)
-# Legendre<--Chebyshev
-function _convert(::Type{Q}, p::P) where {
-    T <: AbstractFloat,
-    Q <: Union{Legendre, OrthonormalLegendre},
-    P <: Union{Chebyshev{T}, OrthonormalChebyshev{T}}}
-    ps = coeffs(p)
-    Q(cheb2leg(ps, normcheb=isorthonormal(P), normleg=isorthonormal(Q)))
-end
+# ## Use FastTransform for conversion, as  possible
+# ## FastTransforms.kind2string.(0:15)
+# # Legendre<--Chebyshev
+# function _convert(::Type{Q}, p::P) where {
+#     T <: AbstractFloat,
+#     Q <: Union{Legendre, OrthonormalLegendre},
+#     P <: Union{Chebyshev{T}, OrthonormalChebyshev{T}}}
+#     ps = coeffs(p)
+#     Q(cheb2leg(ps, normcheb=isorthonormal(P), normleg=isorthonormal(Q)))
+# end
 
-# Chebyshev<--Legendre
-function _convert(::Type{Q}, p::P) where {
-    T <: AbstractFloat,
-    Q <: Union{Chebyshev, OrthonormalChebyshev},
-    P <: Union{Legendre{T}, OrthonormalLegendre{T}}
-}
-    ps =  coeffs(p)
-    Q(leg2cheb(ps, normleg=isorthonormal(P), normcheb=isorthonormal(Q)) )
-end
+# # Chebyshev<--Legendre
+# function _convert(::Type{Q}, p::P) where {
+#     T <: AbstractFloat,
+#     Q <: Union{Chebyshev, OrthonormalChebyshev},
+#     P <: Union{Legendre{T}, OrthonormalLegendre{T}}
+# }
+#     ps =  coeffs(p)
+#     Q(leg2cheb(ps, normleg=isorthonormal(P), normcheb=isorthonormal(Q)) )
+# end
 
-# ultraspherical<--ultraspherical
-function _convert(::Type{Q}, p::P) where {
-    β, α,  T <: AbstractFloat,
-    Q <: Union{Gegenbauer{β}, OrthonormalGegenbauer{β}},
-    P <: Union{Gegenbauer{α, T}, OrthonormalGegenbauer{α, T}}
-}
+# # ultraspherical<--ultraspherical
+# function _convert(::Type{Q}, p::P) where {
+#     β, α,  T <: AbstractFloat,
+#     Q <: Union{Gegenbauer{β}, OrthonormalGegenbauer{β}},
+#     P <: Union{Gegenbauer{α, T}, OrthonormalGegenbauer{α, T}}
+# }
     
-    ps =  coeffs(p)
-    Q( ultra2ultra(ps, α, β, norm1=isorthonormal(P), norm2=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( ultra2ultra(ps, α, β, norm1=isorthonormal(P), norm2=isorthonormal(Q)) )
        
-end
+# end
        
-# Jacobi<--Jacobi
-function _convert(::Type{Q}, p::P) where {
-    γ, δ, α, β,  T <: AbstractFloat,
-    Q <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}},
-    P <: Union{Jacobi{α, β, T}, OrthonormalJacobi{α, β, T}}
-}
+# # Jacobi<--Jacobi
+# function _convert(::Type{Q}, p::P) where {
+#     γ, δ, α, β,  T <: AbstractFloat,
+#     Q <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}},
+#     P <: Union{Jacobi{α, β, T}, OrthonormalJacobi{α, β, T}}
+# }
     
-    ps =  coeffs(p)
-    Q( jac2jac(ps, α, β, γ, δ,  norm1=isorthonormal(P), norm2=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( jac2jac(ps, α, β, γ, δ,  norm1=isorthonormal(P), norm2=isorthonormal(Q)) )
     
-end
+# end
 
-# Laguerre<--Laguerre
-function _convert(::Type{Q}, p::P) where {
-    α, β,  T <: AbstractFloat,
-    Q <: Union{Laguerre{β}, OrthonormalLaguerre{β}},
-    P <: Union{Laguerre{α, T}, OrthonormalLaguerre{α, T}}
-}
+# # Laguerre<--Laguerre
+# function _convert(::Type{Q}, p::P) where {
+#     α, β,  T <: AbstractFloat,
+#     Q <: Union{Laguerre{β}, OrthonormalLaguerre{β}},
+#     P <: Union{Laguerre{α, T}, OrthonormalLaguerre{α, T}}
+# }
 
-    ps =  coeffs(p)
-    Q( lag2lag(ps, α, β,  norm1=isorthonormal(P), norm2=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( lag2lag(ps, α, β,  norm1=isorthonormal(P), norm2=isorthonormal(Q)) )
     
-end
+# end
 
 
-# Jacobi<--ultraspherical
-function _convert(::Type{Q}, p::P) where {
-    γ, δ, α,  T <: AbstractFloat,
-    Q <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}},
-    P <: Union{Gegenbauer{α, T}, OrthonormalGegenbauer{α, T}}
-}
+# # Jacobi<--ultraspherical
+# function _convert(::Type{Q}, p::P) where {
+#     γ, δ, α,  T <: AbstractFloat,
+#     Q <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}},
+#     P <: Union{Gegenbauer{α, T}, OrthonormalGegenbauer{α, T}}
+# }
     
-    ps =  coeffs(p)
-    Q( ultra2jac(ps, α, γ, δ,  normultra=isorthonormal(P), normjac=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( ultra2jac(ps, α, γ, δ,  normultra=isorthonormal(P), normjac=isorthonormal(Q)) )
     
-end
+# end
 
-# ultraspherical<--Jacobi
-function _convert(::Type{Q}, p::P) where {
-    α, γ, δ,  T <: AbstractFloat,
-    Q <: Union{Gegenbauer{α, T}, OrthonormalGegenbauer{α, T}},
-    P <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}}
-}
+# # ultraspherical<--Jacobi
+# function _convert(::Type{Q}, p::P) where {
+#     α, γ, δ,  T <: AbstractFloat,
+#     Q <: Union{Gegenbauer{α, T}, OrthonormalGegenbauer{α, T}},
+#     P <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}}
+# }
     
-    ps =  coeffs(p)
-    Q( jac2ultra(ps,γ, δ, α, normjac=isorthonormal(P), normultra=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( jac2ultra(ps,γ, δ, α, normjac=isorthonormal(P), normultra=isorthonormal(Q)) )
     
-end
+# end
 
-# Jacobi<--Chebyshev
-function _convert(::Type{Q}, p::P) where {
-    γ, δ,  T <: AbstractFloat,
-    Q <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}},
-    P <: Union{Chebyshev{T}, OrthonormalChebyshev{T}}
-}
+# # Jacobi<--Chebyshev
+# function _convert(::Type{Q}, p::P) where {
+#     γ, δ,  T <: AbstractFloat,
+#     Q <: Union{Jacobi{γ, δ}, OrthonormalJacobi{γ, δ}},
+#     P <: Union{Chebyshev{T}, OrthonormalChebyshev{T}}
+# }
     
-    ps =  coeffs(p)
-    Q( cheb2jac(ps,γ, δ, normcheb=isorthonormal(P), normjac=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( cheb2jac(ps,γ, δ, normcheb=isorthonormal(P), normjac=isorthonormal(Q)) )
     
-end
+# end
 
-# Chebyshev<--Jacobi
-function _convert(::Type{Q}, p::P) where {
-    γ, δ,  T <: AbstractFloat,
-    Q <: Union{Chebyshev, OrthonormalChebyshev},
-    P <: Union{Jacobi{γ, δ,T}, OrthonormalJacobi{γ, δ,T}}
-}
+# # Chebyshev<--Jacobi
+# function _convert(::Type{Q}, p::P) where {
+#     γ, δ,  T <: AbstractFloat,
+#     Q <: Union{Chebyshev, OrthonormalChebyshev},
+#     P <: Union{Jacobi{γ, δ,T}, OrthonormalJacobi{γ, δ,T}}
+# }
     
-    ps =  coeffs(p)
-    Q( jac2cheb(ps,γ, δ, normjac = isorthonormal(P), normcheb=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( jac2cheb(ps,γ, δ, normjac = isorthonormal(P), normcheb=isorthonormal(Q)) )
     
-end
+# end
 
-# ultraspherical<--Chebyshev
-function _convert(::Type{Q}, p::P) where {
-    α,  T <: AbstractFloat,
-    Q <: Union{Gegenbauer{α}, OrthonormalGegenbauer{α}},
-    P <: Union{Chebyshev{T}, OrthonormalChebyshev{T}}
-}
+# # ultraspherical<--Chebyshev
+# function _convert(::Type{Q}, p::P) where {
+#     α,  T <: AbstractFloat,
+#     Q <: Union{Gegenbauer{α}, OrthonormalGegenbauer{α}},
+#     P <: Union{Chebyshev{T}, OrthonormalChebyshev{T}}
+# }
 
-    ps =  coeffs(p)
-    Q( cheb2ultra(ps,α, normcheb=isorthonormal(P), normultra=isorthonormal(Q)) )
+#     ps =  coeffs(p)
+#     Q( cheb2ultra(ps,α, normcheb=isorthonormal(P), normultra=isorthonormal(Q)) )
     
-end
+# end
 
-# Chebyshev<--ultraspherical
-function _convert(::Type{Q}, p::P) where {
-    α,  T <: AbstractFloat,
-    Q <: Union{Chebyshev, OrthonormalChebyshev},
-    P <: Union{Gegenbauer{α,T}, OrthonormalGegenbauer{α,T}}
-}
-    ps =  coeffs(p)
-    Q( ultra2cheb(ps,α, normultra=isorthonormal(P), normcheb=isorthonormal(Q)) )
+# # Chebyshev<--ultraspherical
+# function _convert(::Type{Q}, p::P) where {
+#     α,  T <: AbstractFloat,
+#     Q <: Union{Chebyshev, OrthonormalChebyshev},
+#     P <: Union{Gegenbauer{α,T}, OrthonormalGegenbauer{α,T}}
+# }
+#     ps =  coeffs(p)
+#     Q( ultra2cheb(ps,α, normultra=isorthonormal(P), normcheb=isorthonormal(Q)) )
     
-end
+# end
 
 
 ##  Koepf and Schmersa Thm 2, p11 connection for P to Q when σ = σ̂
